+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.11
 
+* Fix: Thread-safe for HaulAIUtility.
 * Compatibility: Thread-safe patch for Zoology: Realistic Animal Overhaul.
 
 ## 1.0.10  (2026-05-05)